Bug 1731214 - Try to fix crash in [@ memcpy_repmovs | nsMsgLineBuffer::BufferInput]. r=kaie
Differential Revision: https://phabricator.services.mozilla.com/D148983

--HG--
extra : rebase_source : b0ba336a450c2ad17e596c60f8ab445d004ed731
extra : amend_source : 19e28971e3eeb3f1d7d92c3e70145f785b27735d

+++ mailnews/base/src/nsMsgLineBuffer.h
@@ -17,7 +17,7 @@ class nsByteArray {
   virtual ~nsByteArray();
   uint32_t GetSize() { return m_bufferSize; }
   uint32_t GetBufferPos() { return m_bufferPos; }
-  nsresult GrowBuffer(uint32_t desired_size, uint32_t quantum = 1024);
+  nsresult GrowBuffer(uint64_t desired_size, uint32_t quantum = 1024);
   nsresult AppendString(const char* string);
   nsresult AppendBuffer(const char* buffer, uint32_t length);
   void ResetWritePos() { m_bufferPos = 0; }
